Updated openstack/openstack Project: openstack/tripleo-heat-templates  0a53ba43cc8a93f47868ffbbeb7faea325529c81
Increase corosync token timeout

This might prevent dropping members from corosync cluster on high load
environments. Symptoms of this problem happening can sometimes be found
in corosync log:

dub 05 17:23:45 overcloud-controller-0 corosync[14152]: [MAIN ] Corosync
main process was not scheduled for 3691.8391 ms (threshold is 1320.0000
ms). Consider token timeout increase.

The default in the Puppet manifest is 1 second, which matches the
corosync default, and we override it with hiera to 10 seconds.

Change-Id: I5ea850ada657e5eecafa3e8b28613a0ac48e78f3