Updated openstack/openstack Project: openstack-infra/tripleo-ci  8553d6f18f09730d14497b9c9c642e778526948e
Wait 20 minutes for tenant stack

We increased the rpc timeout for Heat to 10 minutes in
I1474e8df55e06f4c7e3c9de509f5e38ca727d761
but we were still only waiting 6 minutes with wait_for. This patch bumps
that to 20 minutes. The default timeout for a stack action
(stack_action_timeout) in Heat is 3600s, so we should already be safe
there.

Change-Id: Ic8b4bf12102b6f7f013462706a6c8a71ba428844
Partial-Bug: #1537898