Updated openstack/openstack Project: openstack-infra/project-config  3846cdee0e6856dcc1611697bc8706b2a714a979
Add a job to build openstack-pkg-tools

This patch makes it possible for openstack-pkg-tools to build itself.

The normal package build workflow will be to get openstack-pkg-tools
installed with apt-get, then use pkgos-infra-install-sbuild and
pkgos-infra-build-pkg to build.

Though for openstack-pkg-tools, this can't be used, otherwise we have
a circular dependency, and we expose ourselves to the possibility of
having a bad commit on openstack-pkg-tools to break all of the build
system if a bad commit is published as a .deb.

Which is why openstack-pkg-tools must use its own scripts directly.

Once openstack-pkg-tools is reliably built and its resulting .deb
published on an OpenStack infra Debian repository, then a new generic
package build job will be added to project-config.

Change-Id: I7a13b2f298b5426219b6d3842ac1a9cf7a2a3656