Updated openstack/openstack Project: openstack/cinder  557ff9297957d8a8ac20ebd02ac8b6915b3d0c07
Don't disallow quota deletion if allocated < 0

Due to the race condition http://bugs.launchpad.net/cinder/+bug/1552944,
it's possible that a volume delete on a project at the same time as a
change to/from a -1 limit on the project could double count the change
in allocated value. This has the possibility to set allocated < 0, and
in order to handle this situation as gracefully as possible, we should
not disallow quota deletion in the case that allocated < 0.

Change-Id: Ic35ff9671024aa1cecc04b205d1b7cba56bbea61
Closes-Bug: #1554173