Updated openstack/openstack Project: openstack-infra/project-config  a59e02278c4bb679b78ea05631f29c7f56ffebcc
Add post job for building and uploading tinyipa

Ironic currently automatically builds and uploads a coreos based IPA
image so that CI jobs don't have to build it every time and so that
people wanting to use Ironic don't have to always build their own. This
patch adds a job to also build the tinyipa version of IPA, so that we
can get the same optimisations.

Depends-On: I9d879689c8e88768aaa96d0033b62d062f9cdc27
Change-Id: I56688bbb3efd5b5a0a14570f4b9733a76d31b71b