Updated openstack/openstack Project: openstack/keystone  ad7a7bd6ee36a7af61f88d98038d83aba25a9743
Remove backend interface and common code out of identity.core

This removes dependencies where backend code references code in the
core.  The reasoning being that the core should know about the backend
interface, but the backends should not know anything about the core
(separation of concerns).  And part of the risk here is a potential for
circular dependencies.

- Moved the filter_user method to backends/base.py.  This method is
only used by the backends.

- Moved the interfaces (base) to modules inside the backends. Interfaces
should be closer to the code that implements them and again, wanted to
remove the dependency where the backends were referencing code in the
core.

Partial-Bug: #1563101

Change-Id: I5ff9c4e4b6d64750f5db2a73cc4317358aea0649