Updated openstack/openstack Project: openstack/tempest  19a29d80bf7302295de53d5e30c1ebca09802090
Skip test_compare_db_flavors_with_os

This commit skips test_compare_db_flavors_with_os. This has a 100%
failure rate in the gate right now. This is likely related to the
devstack nova change creating the default flavors. But, if so this
indicates how fragile the test is and it's unrealistic dependence
on the old defaults. Let's skip for now to unblock things.

Related-Bug: #1567134

Change-Id: I6e3ef31f5bff42e4ff80742e4649ea667dc945b5