Updated openstack/openstack Project: openstack-infra/project-config  72d2545763cbfd1b0855976f0cad2a1d843a0961
Reordered builders in npm builds

Some of our npm builds are breaking because the cache has not yet
been updated by the time the npm-install macro runs. This patch
does two things - ensures that apt-get update is invoked before
prerequisites install, and reorder the builders in our javascript
templates to match that of the other jobs.

Change-Id: I44440c9da33eeb86d703ab4e605c31b02df1335c