Updated openstack/openstack Project: openstack/neutron  ba7c7bc0a92204e8e970ccd05cad0bc7a3a00963
QoS DSCP use mod_flow instead of mod_flows

In the implementation of DSCP QoS rule, the QosOVSAgentDriver uses the
wrong method to modify br-int flows. It uses br_int.mod_flows() whilst
it should use br_int.mod_flow().

This patch fixes this and also adds verification of the updates of DSCP,
as we have for bandwidth here to trigger that code path and avoid
regressions.

Change-Id: I685ac373701ff8407fd7fbf649e17a2f7dfc0008
Closes-Bug: #1564820