Updated openstack/openstack Project: openstack/neutron  1efed3a5321a259d27ec4a6e80352d35fc423c11
Fix race conditions in IP availability API tests

DHCP port creation is asynchronous with subnet creation.
Therefore there is a time window where, depending on how fast
the DHCP agent honors the request, the DHCP port IP allocation
may or may not be accounted for in the total number of used IP
for the network. To kill the race, do not run dhcp on the
created subnets at all.

Closes-bug: 1563883

Change-Id: Idda25e65d04852d68a3c160cc9deefdb4ee82dcd