Updated openstack/openstack Project: openstack/designate  ccf53f48c3bb7f7a499321862a55f8699c737d8c
Merge tag '2.0.0'

This is a null-merge of the 2.0.0 release tag back into the master
branch so that the 2.0.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: Icea749639a2f1d8cb87950646fbff32b83551a72