Updated openstack/openstack Project: openstack/horizon  05d49af8e6c53378b298114331f1195cc7c82e1d
Prevent 'Detach Interface' action if an interface is not attached

In the Instance action menu of Instance panel, 'Detach Interface' action
need not be displayed if the instance does not have an attached interface.
A check has been added to validate if there are any 'fixed' IP address is
associated with the instance, to enable 'Detach Interface' action menu.

Change-Id: I73d4c052b4aa12a50887c220d1bcd3a0b3f9a44d
Closes-Bug: 1506432