Updated openstack/openstack Project: openstack/oslo.config  14ec2c67eb2a9720b0cf567d3db6174626b594c9
Add Port type to allow configuration of a list of tcp/ip ports

There was already a PortOpt, but this is not composable with ListOpt.
The new type inherits from Integer and therefore supports min/max
value and choices. For example, openstack/vmware-nsx wants to use this
to configure a default list of ports for a firewall rule.

Change-Id: I83bbec6add8ce2951e94e69ec14bb6d8137d4f0c