Updated openstack/openstack Project: openstack/nova  e2778f651371090ad81b3348e22ac75d3d492942
Merge tag '13.0.0'

This is a null-merge of the 13.0.0 release tag back into the master
branch so that the 13.0.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: I8fb95088a8269a28734d6582d5ca1f9472682a6e