Updated openstack/openstack Project: openstack/tripleo-heat-templates  706c2fe4b62f95ac13ee800fc08e549180afc810
Add removal of the /etc/resolv.conf.save file for +bug/1567004

The change at https://review.openstack.org/#/c/302352/ should stop
the if up/down scripts from making changes to resolv.conf as
discussed in that review and the related bug below. However during
upgrades, as we are moving from a version of the ifcfg-vlanXX files
that don't have the PEERDNS=no added by /#/c/302352 the if up
script will restore the /etc/resolv.conf.save to /etc/resolv.conf
and overwrite it. This removes the .save file during the upgrade
init command which gets delivered to all nodes as the first stage
of a major upgrade.

Change-Id: I91dd139f43be4912c20d8661691bee2b662964d4
Related-Bug: 1567004