Updated openstack/openstack Project: openstack-infra/shade  8601dd68f7e4949e94f9e6074657b0e36a146855
Update func tests for latest devstack flavors

The most recent devstack adds new flavors to the existing set, so
we need to adjust some functional tests that are testing range
searches across flavors.

Note that we have a gate test that tests against stable/liberty
and that devstack still has the original m1 flavors only. So we
need to account for that in the functional tests.

Change-Id: I98ce866686f3243ef2d2800b872a9840761fbe1a