Updated openstack/openstack Project: openstack/nova  fce812c76f350ebe2113ec2e028c251e8e3d2696
Add functional test for OverQuota

Since we now have checks for OverLimit on both volume and snapshot
creation, we need to check and make sure OverQuota is raised properly
for both volumes and snapshots. Because snapshot create was not checking
for OverQuota, that needed to be caught in the API so it could be
re-raised properly.

Helper methods for snapshot operations in the functional test API client
were missing, so this also added those in so we can use them in the
functional test.

Change-Id: Iea16a6cb7d3448e0bdd8491826b0780fa8e1a9b1
Closes-Bug: #1554631