Updated openstack/openstack Project: openstack/cinder  3996dc8a67c643fa1e793ef53136de311d998b4d
Merge tag '8.0.0'

This is a null-merge of the 8.0.0 release tag back into the master
branch so that the 8.0.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: I0b8fef4b44eb8375daf361f676997ae83347b5a0