Updated openstack/openstack Project: openstack/ceilometer  fa5711c1a0294c12fa80a070c24b66d3ad38dd42
Merge tag '6.0.0'

This is a null-merge of the 6.0.0 release tag back into the master
branch so that the 6.0.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: I74789b3e18ef6b478a55615a9c4b7ea9200f8e5c