Updated openstack/openstack Project: openstack/oslo.messaging  5e0e0e09013b6c50e236996fca41a726c916d087
Allow simulator to be launched from arbitrary directory

Right now simulator.py expects messages_length.yaml to be in the
current directory and as a result, it is impossible to launch
simulator while you are at the repository root.

The change makes simulator to search for the file in the same
directory as simulator itself is in.

Change-Id: Ic2ca21d87e2d888432dc6204c7b4b995a06aecbd