Updated openstack/openstack Project: openstack/glance  128a31701ed7b72effc601334bef7dc6f3f2c8d2
Merge tag '12.0.0'

This is a null-merge of the 12.0.0 release tag back into the master
branch so that the 12.0.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: Ide9dbf6e5b83dd7f6d7dfcbe49968603019c4409