Updated openstack/openstack Project: openstack/ceilometer  11a1a385b1565115df4fa57b7402fdedd8cd041e
gnocchi: batch measurements

This change leverages the Gnocchi Batch API by using it
in the Ceilometer dispatcher for measurements.

It introduces a small behavior change in the dispatcher.
The resource metadata update is now decorrelated from measures posting.
So, even if posting measurements fail, the resource metadata will
be updated.

Change-Id: I140e7301a31bbee50c2c8bc3ff2d78925680a573