Updated openstack/openstack Project: openstack/glance_store  871d34d393094d9e97ad1c9c538c4b7fe7eaeaab
Deprecate swift driver options properly

Some swift driver options were deprecated some time ago but it
turns out that users are not aware of that. In order to notify
users about deprecation and not breaking their envs after
option removal we need to use deprecation flags in configuration
options. So it must be shown in log files.

Change-Id: I083eb6f00324b3a4b9a3313497993beab776cc52
Related-Bug: #1561947