Updated openstack/openstack Project: openstack-infra/project-config  ce6b03963143e141f545976ab3df02e9a20696e4
Make gate-neutron-dsvm-api job non-voting

This is done as a temporary measure until we can drastically limit
the string on failures that have been hitting since change a2fe5f.

Initially a similar attempt was done with [1], but that strategy
was not deemed worthy. Whilst we work out how to complete the
decoupling effort, it is safer to allow the wider team to work
without too much disruption.

[1] https://review.openstack.org/#/c/274182/

Change-Id: I3ee94e31e091f6ac35b2503f960720cbe49ac548