Updated openstack/openstack Project: openstack/sahara  2582847397f0cddc648a5b875199f0d428613ddc
SPI Method to Validate Images

Adds SPI method and utility module to Sahara. Does not yet include any
yaml files or plugin implementations in this change set; early
implementations are in the works for testing but not yet ready for
review, and will come in separate commits for modularity.

Partially-implements: blueprint validate-image-spi
Change-Id: I9f59219a801325176a70fd8da1ca3b38efc5b109