Updated openstack/openstack Project: openstack/nova  e6aab744a38395ccbf8faaee80085bd4765edabe
Add metadata objects for device tagging

This patch series adds the ability for a user to specify a tag to be
applied to a network interface or a disk when booting an instance.
This tag is then exposed through the metadata API. For example, a user
boots a VM with two network interfaces. One is connected to a private
network, the other to the public Internet. There is currently no
direct way to tell which interface is which. Specifying tags allows
the user to distinguish between the two interfaces.

This patch introduces the new metadata objects.

Implements: blueprint virt-device-role-tagging
Co-authored-by: Vladik Romanovsky <vromanso@redhat.com>
Change-Id: I8e7100f83ef8504f19b156283e0ade0fcb37dc56