Updated openstack/openstack Project: openstack/nova  a9459d3c41fa28dbbdac01d058d4a45e78906f7c
remove alembic from requirements.txt

Alembic was used in attempt #1 of online schema migrations, however
that was reverted in Icae28ceee3ec975c907d73b95babab58dcb30c23 when
that approach was dropped.

There are no other uses of alembic directly in Nova, so we should not
list this requirement.

Change-Id: I452bfc8454aedff1bbaffacc99d0845186ba4234