Updated openstack/openstack Project: openstack/barbican  56c82cedeff3b11e808514b816085b511d5aee40
Handling json-home header for /v1 call

As per related bug, currently seeing CRITICAL failure in barbican logs
which raises unnecessary alarm if someone is monitoring logs for CRITICAL
issues. The error is related to unexpected client input and code is not
able to determine renderer for json-home input case. So just setting
override template when version response is built.

Change-Id: I0f6e5d19969523ddacef6e2bf0c70030af2d049b
Closes-Bug: #1568141