Updated openstack/openstack Project: openstack/cinder  a093a3558e3fd0abd86f38c69823e7d283ee02bf
EMC VMAX - Operations and timeout issues

If a volume operation e.g create, delete, expand, etc were to
timeout without fully completing then we were not appropriately
catching the error with the appropriate rc (return code) or
errordesc (Error description). This new method does an additional
check to appropriately capture the status of the job i.e. did it
fail to complete in the allocated time.

Change-Id: I0b74f9941d722083b70c12bbebeb269d315c02b3
Closes-Bug: #1538214