Updated openstack/openstack Project: openstack/cinder  7811cb8d13d59f94d3dd16a4a2abc0f675bcaa9e
Fix meaningless quota exceeded exception info

When updating a nested quota with exceeded value, a meaningless
info is returned to user.

In Cinder, exception's attribute 'message' is a format string,
not including details. We need to use 'msg' instead of it.

See:
https://github.com/openstack/cinder/blob/master/cinder/exception.py#L122-L126

Thanks to ChangBo Guo(gcb) who find this:
https://review.openstack.org/#/c/285712/

Change-Id: I99c891b865ec507875d2a48e3e7d1f4a3b1bb9cb
Closes-bug: 1561413