Updated openstack/openstack Project: openstack/nova  bbb5373ebc95a48c663abc98aade2a6e404b06cc
Improve the help text for the linuxnet options (1)

The previous commit moved all the options from network/linuxnet.py to a
central location in nova/conf/network.py. This commit adds additional
help text to those options.  Because of the large number of options, the
improved help text will be split into several patches; this is patch 1
of 4.

Blueprint centralize-config-options-newton

Change-Id: I79b587aa6bba902f53c786646f47f610ca9bc2fd