Updated openstack/openstack Project: openstack/nova  d7c9cb07d5bc6b8bba4c68b1bfc005b3df3750e7
Return 400 on boot for invalid image metadata

Horizon allows a free-form input field for architecture
when creating an image. However, the architecture in
the image metadata is validated against nova.compute.arch.ALL
so when converting the image metadata in the compute API
boot validation, if we hit a ValueError we should raise
an appropriate exception so the REST API will return a 400
rather than a 500.

Change-Id: Ib01b4b088fec065c671f83f5912e928b4d7fec71
Closes-Bug: #1558866