Updated openstack/openstack Project: openstack-infra/project-config  863cdac470527460db388985e3c4af1ddf9645f8
Mutlinode fixup - let Watcher plugin install CLI

The Watcher multinode experimental job is currently not working due
to the fact that the DEVSTACK_PROJECT_FROM_GIT variable is actually
shared by both the primary node and the subnodes. Therefore, I removed
it from the multinode job to let the Watcher plugin deal with this
using 'is_service_enabled' to detect whether we are on the primary
node or not.

Change-Id: I292d109c89c9f07b1440a215d6c50b9243df9abe