Updated openstack/openstack Project: openstack/barbican  f428dd6be9da4e00dcc01719c3c3d7454d58505a
Add skips for KMIP functional tests

In preparation for adding a KMIP gate, adds the ability to skip functional
tests that are not compatible with a KMIP device. Some tests are skipped
because invalid algorithms or bit lengths are passed to the device. Other
tests are skipped because the tested feature doesn't exist for the PyKMIP
software server yet.

To enable skipping the non-compatible functional tests for a KMIP back
end, set the environment variable using:
    export KMIP_PLUGIN_ENABLED=1

Co-authored-by: Nathan Reller <nathan.reller@jhuapl.edu>
Change-Id: Id908bf57233af84bff56d90c75d175b04ccd4373