Updated openstack/openstack Project: openstack/nova  8991c27b21bdba890b7002b55c17fdb414839ca5
Add link to the latest nova.conf example

The current README-nova.conf.txt implies that you need to clone the
entire repo and build the sample configuration file yourself. While that
is a viable option for some, we should also point out that you can see
an example of the latest nova.conf on docs.openstack.org, which
obviously has a much lower barrier to entry.

Change-Id: I94aa183c64287549d30bf09b01c3c3f7fa86be27