Updated openstack/openstack Project: openstack/ironic  157282b83e7afa00dbfa84f2349cf3af9ec6f376
Rename juno name state modification method

The provision name state modification method exiplicitly has juno
in the name of the method. Since this is a method to convert the
returned states to clients in backwards compatible ways and
it will be present until the end of time, we should name it
something that does not have a connotation with a specific
release.

Change-Id: Ie3c7c2b1e7557f9187885f623e12ab0da2d63bbf