Updated openstack/openstack Project: openstack-infra/project-config  81d8c8449f9958fb36ad0ec98f73805ab914fed2
Fix Babel to 2.2.0 for now for django translation

Looking at today's django imports for horizon like
https://review.openstack.org/#/c/303695 and
https://review.openstack.org/303807 , a lot of strings are removed
from pot files - which looks wrong. There was just a Babel 2.3.2
release, let's go back to 2.2.0 that does not remove these strings.

We need some more time to investigate this and fixing the version to
2.2.0 will give us this time. I fear the translated strings are already
removed from zanata but we might be able to get them back from git.

Change-Id: I952d842e4406b33257e8d715b7ac774a5f6c4f4f