Updated openstack/openstack Project: openstack/oslo.messaging  c81aa0283225798e0186d9f8c6604b7b03e6be7a
[kafka] Do not remove kafka_client during reset

Currently we delete kafka_client, producer and consumer
from the Kafka driver connection when we reset it,
for example before returning it to pool.

It's a redundant action, because kafka_client and
kafka producer (it uses a kafka_client for
sending a message) could be used again without any
concerns.
In same time, currently we don't close a KafkaConsumer,
but it's needed because we should to close opened
sockets to the kafka.

In this patchset all action in reset are changed
to the more optimal behavior.

Change-Id: I6ff26256c933c79468f9e6cd1752181e5ace155f
Closes-bug: #1557528