Updated openstack/openstack Project: openstack/python-openstackclient  cb8f757779ded16b769bce2c990650662dfa5223
Refactor security group rule list to use SDK

Refactored the 'os security group rule list' command to use the
SDK when neutron is enabled, but continue to use the nova client
when nova network is enabled.

In addition, a release note was added to document the features
and fixes resulting from this refactor.

Change-Id: I24d04b720102ed1c60776e1ca67d4ca20e31b663
Partial-Bug: #1519512
Implements: blueprint neutron-client