Updated openstack/openstack Project: openstack-infra/shade  a26a9bcb6083784cdf4e11efc64c5544dd5b9a48
Use event_utils.poll_for_events for stack polling

Calling get_stack to poll for stack state transitions causes
unnecessary high load on heat servers so should be avoided if possible
(this is true whether get_stack lists all stacks or a fetches a single
full stack).

Heatclient has a utility function which instead polls for stack events
(with a fallback to fetching the stack when events are not
forthcoming).  This function is used extensively by heat client and
the openstackclient stack commands - it would be appropriate to use it
here too.

The timeout is passed to the stack create call, meaning that the stack
will go to CREATE_FAILED if the timeout is exceeded. The default
timeout_mins is usually 60 minutes, so the client-side timeout would
never be reached anyway.

Also, the current polling approach was not filtering for
CREATE_COMPLETE so it wasn't actually waiting for anything.

This change adds functional tests which cover get_stack, create_stack
and list_stacks. test_stack_nested exercises the stack_create
environment_files file composition.

Change-Id: Ia14d47f0f51e1f8825b6de6d8dc5a12335913f55