Updated openstack/openstack Project: openstack/oslo.utils  a99969ef8a0ce7232fab10807ee1b46dcb5817f4
Provide single step check if eventlet is monkey_patched

Currently to perform such check we need to check if eventlet
is available first, that give us two steps, but it is
desired to check this at once.

This change is to remove the function here:
[1] https://github.com/openstack/oslo.messaging/blob/master/oslo_messaging/_drivers/pika_driver/pika_commons.py#L37
Also will be useful in zmq-driver.

Change-Id: Ie24ecb54946bfba465d293185ebfda25691a8701