Updated openstack/openstack Project: openstack-infra/project-config  3302b3ea66678252165a4ac83d11f9ed1c771ec1
Add js-afs-blob-store (afs-blob-store) to openstack-infra

This project is an implementation of the fs-blob-store, which exposes
an API that appears to be a real POSIX filesystem, but in reality
is a file renaming shim which sorts all documents into alphabetical
directories. /foo becomes /f/foo, /bar becomes /b/bar, etc.

This project backs our NPM mirroring script, allowing us to mirror
npm into AFS.

Needed-by: Icbaa3d28650583599d20f273ef65757783a73674
Change-Id: Ic969dd933349f718409dc9357b25ff13c0e02d29