Updated openstack/openstack Project: openstack/nova  8d28e2a137eae64d07e10af00668b7e0314d2319
Fix max concurrent builds's unlimited semaphore

When booting an instance with the config paramter
max_concurrent_builds set to 0, a TypeError is logged in the compute
log. This doesn't prevent the instance from booting, but could be
confusing to someone debugging a different issue. This patch fixes
this issue by fixing the method signature of the __exit__ method of
compute utils's UnlimitedSemaphore. UnlimitedSemaphore was bypassed
entirely in unit tests because the default value for
max_concurrent_builds is 10.

Closes-bug: 1520633
Change-Id: I15825eb3fb1bccb12f4a9148809495ce3e83e3c8