Updated openstack/openstack Project: openstack/ironic-python-agent  819b79da87c01b2af49ef4d1871065ce7aa768ee
Add script to install missing tinyipa dependencies

Squashfs-tools is missing when building the tinyipa image during the
post job, this patch adds a new script to the tinyipa build process that
ensures that missing dependencies get installed as part of the build
process. It also replaces a use of realpath it build-iso.sh where it
should be readlink to ensure that we don't need that dependency.

Change-Id: I9a60dcf42f6a9ba624aa347f0f375bb40432ef77
Closes-Bug: #1567981