Updated openstack/openstack Project: openstack/keystonemiddleware  0f426e7f9e86385cab698f73d6ecb1d8e6b4898a
Create signing_dir upon first usage

Without this patch, if a signing_dir value is not provided in the
keystone_authtoken section for a given service, keystonemiddleware will
create a secure temporary directory upon startup. For users not using
the 'pki' pr 'pkiz' token providers, this is unnecessary and causes
/tmp to fill up needlessly. This patch moves the directory
creation out of SigningDirectory.__init__ and into
SigningDirectory.calc_path, so that the first time calc_path is called,
the directory is created. If the user has provided the directory name,
we still log that and verify the directory at __init__. If the temp
directory must be created, we log and verify that only once at the time
of creation.

Change-Id: I501d96d35bcf0039d2f5c5987471f496fc830f61
Closes-bug: #1533724