Updated openstack/openstack Project: openstack/nova  95ede914f81309bb23e4fe646264d10b9c09d31a
Migrate compute node resource information to Inventory objects

This does a live migration of compute node resource information from the
current location (in compute_nodes columns) to the new desired location
of records in the inventories table. It only does this if all control
services have been upgraded to a level that understands this.

Related to blueprint compute-node-inventory-newton

Change-Id: Ieda099d3f617713f09315b2b1e932a7c1d6f45c4